refactor(analysis): split content function span detectors
